pyproject.toml
[project]
name = "examen"
version = "0.1.0"
description = "Sistema de Florería - Ventas de Flores y Plantas"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-login>=0.6.3",
    "flask-admin>=1.6.1",
    "flask-migrate>=4.0.5",
    "flask-sqlalchemy>=3.1.1",
    "pydantic>=2.5.0",
    "pymysql>=1.1.0",
    "cryptography>=41.0.0",
    "python-dotenv>=1.0.0",
    "werkzeug>=3.0.0",
    "wtforms>=3.1.0",
    "email-validator>=2.1.0",
    "pillow>=10.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-flask>=1.3.0",
    "black>=23.12.0",
    "flake8>=7.0.0",
]
